xyz/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic>=2.9.0
python-dotenv==1.0.1
google-cloud-aiplatform==1.142.0
anthropic==0.34.2
chromadb==0.5.5
requests==2.32.3
